Move `CreateBoxToPageAssignments` command
